refactor features folder
