+++ src/features/task/components/TaskViewSwitcher.tsx
@@ -12,8 +12,8 @@ import { fetchTasksService } from "@/services/taskService";
 import { useWorkspaceId } from "@/features/workspace/hook/use-workspace-id";
 import { useProjectId } from "@/features/project/hook/use-project-id";
 import { Loader } from "lucide-react";
-import { DataKanban } from "./data-kanban";
-import { DataCalendar } from "./data-calendar";
+import { DataKanban } from "./kaban/data-kanban";
+import { DataCalendar } from "./calendar/data-calendar";
 import { useCallback } from "react";
 import { TaskStatus } from "@/types/task";
 import { updateMultipleTasks } from "@/services/taskService";
