+++ src/features/task/components/calendar/data-calendar.tsx
@@ -12,7 +12,7 @@ import { useState } from "react";
 import { Calendar, dateFnsLocalizer } from "react-big-calendar";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import "./data-calendar.css";
-import { EventCard } from "./event-card";
+import { EventCard } from "../event-card";
 import { CalendarIcon, ChevronLeftIcon, ChevronRightIcon } from "lucide-react";
 import { Button } from "@/components/ui/button"; // Assuming you have a Button component
