+++ src/features/task/components/kaban/data-kanban.tsx
@@ -5,8 +5,8 @@ import {
   Droppable,
   type DropResult,
 } from "@hello-pangea/dnd";
-import { ResponseTask, TaskStatus } from "../../../types/task";
-import { KanbanColumnHeader } from "./kanban-column-header";
+import { ResponseTask,TaskStatus } from "@/types/task";
+import { KanbanColumnHeader } from "../kanban-column-header";
 import { KanbanCard } from "./kanban-card";
 import { useQueryClient } from "@tanstack/react-query";
 const boards: TaskStatus[] = [
